requirements.txt
pyyaml==6.0.2
python-dotenv==1.0.1
openrgb-python==0.3.1
